+++ app/controllers/welcome_controller.rb
@@ -1,0 +1,4 @@
+class WelcomeController < ApplicationController
+  def version
+  end
+end