+++ config/routes.rb
@@ -1,4 +1,5 @@
 Rails.application.routes.draw do
+
   resources :books do
     member do
       post :checkin
@@ -9,6 +10,8 @@ Rails.application.routes.draw do
     end
 
   end
+
+  get '/version' => 'welcome#version', as: :version
 
   # The priority is based upon order of creation: first created -> highest priority.
   # See how all your routes lay out with "rake routes".