+++ test/controllers/welcome_controller_test.rb
@@ -1,0 +1,10 @@
+require 'test_helper'
+
+class WelcomeControllerTest < ActionController::TestCase
+
+  test "should get version" do
+    get :version
+    assert_response :success
+  end
+
+end